wip: moved ants print functions to another file
